Custom hook for PersonScreen
